change id datatypes to long

+++ user_lib/src/main/java/com/crumbs/lib/repository/DriverRepository.java
@@ -4,5 +4,5 @@ package com.crumbs.lib.repository;
 import com.crumbs.lib.entity.Driver;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DriverRepository extends JpaRepository<Driver, Integer> {
+public interface DriverRepository extends JpaRepository<Driver, Long> {
 }
